Misc: Fix most level 3 warnings

+++ src/tools/ToolsCataMop/map_extractor/System.cpp
@@ -395,7 +395,7 @@ uint32 ReadMapDBC()
 
     SFileCloseFile(dbcFile);
     printf("Done! (%u maps loaded)\n", uint32(map_count));
-    return map_count;
+    return static_cast<uint32_t>(map_count);
 }
 
 #define SZFMTD "%" PRIuPTR
